Throw an error if the given dataurl has a duplicate parameter key
The spec is not clear about this. However, if this case happens, there
is no reason to arbitrarily keep the last parameter only (this was the
previous behavior) or the first only. I believe it's better to be clear
with the user and throw an error for this very specific case.